codon_harmony/codon_harmony_dir/requirements_dev.txt
pip==19.0.3
bumpversion==0.5.3
wheel==0.33.1
tox==3.7.0
Sphinx==1.8.4
sphinx-argparse==0.2.5
coverage==4.5.2
codecov==2.0.15
biopython==1.73
python_codon_tables==0.1.8
